+++ lib/grape_oas/api_model/parameter.rb
@@ -8,9 +8,10 @@ module GrapeOAS
     # @see https://swagger.io/specification/
     # @see GrapeOAS::ApiModel::Operation
     class Parameter < Node
-      attr_accessor :location, :name, :required, :description, :schema, :collection_format
+      attr_accessor :location, :name, :required, :description, :schema, :collection_format, :style, :explode
 
-      def initialize(location:, name:, schema:, required: false, description: nil, collection_format: nil)
+      def initialize(location:, name:, schema:, required: false, description: nil, collection_format: nil, style: nil,
+                     explode: nil)
         super()
         @location = location.to_s
         @name = name
@@ -18,6 +19,8 @@ module GrapeOAS
         @schema = schema
         @description = description
         @collection_format = collection_format
+        @style = style&.to_s
+        @explode = explode
       end
     end
   end